+++ services/server/src/server/apiv1/verification/private/stateless/private.stateless.handlers.ts
@@ -266,7 +266,7 @@ export async function replaceContract(
       const transactionHashFromDatabase = (sourcifyChain as SourcifyChainMock)
         .contractDeployment?.transaction_hash;
       if (transactionHashFromDatabase) {
-        transactionHash = `0x${transactionHashFromDatabase}`;
+        transactionHash = transactionHashFromDatabase;
       }
     } else {
       // Use the chainRepository to get the sourcifyChain object and fetch the contract's information from the RPC
@@ -315,14 +315,22 @@ export async function replaceContract(
       throw error;
     }
 
+    let rpcFailedFetchingCreationBytecode = false;
+    try {
+      rpcFailedFetchingCreationBytecode =
+        verification.onchainCreationBytecode === undefined;
+    } catch (error) {
+      // verification.onchainCreationBytecode throws if not available
+      rpcFailedFetchingCreationBytecode = true;
+    }
+
     res.send({
       replaced: true,
       address: address,
       chainId: chainId,
       transactionHash: transactionHash,
       newStatus: verificationStatus,
-      rpcFailedFetchingCreationBytecode:
-        verification.onchainCreationBytecode === undefined,
+      rpcFailedFetchingCreationBytecode,
     });
   } catch (error: any) {
     throw new InternalServerError(error.message);